feat(item): 상품 검색 페이지네이션 - 가격 범위, 상품 이름

+++ src/main/java/com/example/market/shop/dto/SearchItemDto.java
@@ -15,6 +15,6 @@ public class SearchItemDto {
     private String name; // 상품 이름
     private ItemCategory itemCategory;
     private ItemSubCategory itemSubCategory;
-    private Long minPrice;
-    private Long maxPrice;
+    private Integer minPrice;
+    private Integer maxPrice;
 }
